Add a workaround for a bug in clang-tidy 16 and C++20.
In that mode, the Bazel suppression of a Clang warning doesn't actually
suppress the clang-tidy check. Neither does disabling the check in
a `.clang-tidy` config file. Instead, disable it directly on the command
line when running `clang-tidy`. The clang-tidy bug is:
https://github.com/llvm/llvm-project/issues/61969

This is harmless for other versions as this check is not useful with
most Bazel configurations, and can be achieved with a compiler warning
if truly needed.

Fixes #29