Add clang_tidy_test
This is a new test rule that executes clang-tidy as a test via shell
script instead of an aspect.

The biggest implementation issue with this is that the paths you get
from CcInfo do not match the directory structure when running clang-tidy
in a test. This requires a bit of fixing up.

This attempts to share as much logic as possible with the aspect. I
think a bit more could be shared in the future.

Based on https://github.com/erenon/bazel_clang_tidy/pull/65

Closes: https://github.com/erenon/bazel_clang_tidy/pull/65
Fixes: https://github.com/erenon/bazel_clang_tidy/issues/15